backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# AI/ML
scikit-learn==1.4.0
xgboost==2.0.3
shap==0.44.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# HTTP Client
httpx==0.26.0

# SMS
twilio==9.0.0

# Utilities
email-validator==2.1.0
fpdf==1.7.2
